added 2.4, made 42.3 html data into actual html,..
I added 2.4, with just a table of computer parts for the assignment. I
need to finish 42.3 and 31.3, and resubmit 31.2 before I can continue
with this one, though.

The *.mht generated by Excel was also a mess (it is an *.mht, I guess),
so I changed it into standard HTML5. I removed the original as a result.

In index.php, I worked on improving the include code, and the conversion
of markdown emphasis, boldening and code in block elements to HTML.
Regex look-arounds came into this, and I now like regex even more. 42.3
has changed as a result of fiddling with including (which is actually
proken ATM).