css fixes, docs update, 6 images for 31.2, reorg..
I added more css to the stylesheet, handling the line-height on blocks
of code saparated by <br>s instead of actual newlines. This wouldn't be
nearly as much of a pain in the rear is Windows used good line endings.
Change to \n, Microsoft, and leave Apple in the dirt.

The documentation index has chnaged to show the suggested nginx
congifuration as a included file instead. The macro for 42.3 was also
split into a new file.

I added 6 storyboard images for 31.3, as PNGs, and they have been
optimised. They were created with Cinema 4D using the cel render effect
to make it look storyboard-ish.

I created a new file in the inlcudes directory (which was previously
"include"), which will become a syntax high-lighting class. This will be
very basic, if I were to hazard a guess, and may be replaced in time
with someone else's library/script.

index.php has, as usual, changed for some reason. Probably more
finicking with regexes.

I modified the three unit 42 documents to update file download links,
and added some more to the technical documentation.