changed div.i and object styles, 42.4, 31.3 tweaks
I added styling to the main stylesheet to handle Flash objects a little
better and realised that imaged were in fact meant to be separated by 8
pixels, not 16. The result it a little iffy but it's better than it was.

Work has been done on the new 42.4, completing about half of the pass task.
As part of this I added two images, saved from Photoshop with only 64
colours (should be barely compressible).

I also tweaked 31.3 in some way or another. This was probably a small
syntax change, as the infrastructure and directory heirarchy around the
markdown keeps changing.