changed external anchor styles, added 31.3 fla,..
I have chnaged the stylesheet so that the "ext" class must be added to
an anchor for the external link icon to show after it. This is different
to how it was before, as it used to require the "ref" class in oder to
NOT show up. I have et to modify older assignments markdown documents to
change to this format, but I will do this soon and implement an
automatic class adder in index.php.

I added the Flash project file for 31.3, as opposed to just the SWF. I
also removed the GIF version of the stick-figure animation entirely, as
the second animation is in GIF form an it otherwise only hurts
performance and page load.

31.2 and 31.3 have both been modified. I think the changes to 31.2 are
minimal tweaks. 31.3 is due late tomorrow, and requires feedback from peers
and has a distinction task.