refixed 31.2, continued with 31.3
Near the end of 31.2 is an image div with another div inside it,
allowing for the roll-over image-change, but apparently parsedown can't
handle nested divs and borks out. This was fixed by putting the whole
div.i on one line. I actually had fied this beofre, I think, by removing
the indentation, but then stupidly undid my changes thinking there was a
mistake.

31.3 has been continued, by properly starting the distinction task.
There is more o come, but I was thankfully given an extension to about
this time tomorrow. 31.3 and 42.4 must both be complete by this time
tomorrow.