added 1.1, table styling, css comments, fix,..
I added the 1.1 assignment doc, with some content. This is only a pass
task and should be completed tonight, but it realistically won't be.

I changed the styling for tables, altering the vertical alignment of text
in heading and body sections.

The heading handler class has been tweaked by changing a 2 to a 1. This
fixed headings from all being one size too small.