renamed m.css, finished 1.1 (hwat?)
I have fallen to the gods of neatness over minification, slightly. I have
renamed m.css to main.css, as it looked silly sat next to mobile.css and
highlight-vba.css. I suppose this fits with my nefwound love for for more
blank lines than is socially acceptable in the PHP scripts. As many a
developer can tell you, verboseness may not look pretty or be efficient,
but it is maintainable, which matters infinitely more in the real world.
After a valiant struggle with my own stubborn self, I committed (heh) to
forcing viewers of this project online to download an extra three
bytes of data. Ruthless, right?

In other news, I have bizarrely finished 1.1, much to my own surprise.
It's been proofread and converted to a word document, but it remains to
be uploaded, thanks to the incompetence of British Telecom.

As seems to the be norm now, index.php has changed somehow. Hopefully it
isn't Linux converting the whole back to Unix line-endings again.