1.1 work (always more to do), style changes,..
I have worked on 1.1, as more is apparently needed on that front. Not
the end of the world, I suppose. More is still needed, although not
loads.

The heading handler lass has been changed a bit, as I noticed the
headings on the 1.1 assignment weren't getting indexed properly. This is
still not fixed, so I shall have to look into this some time.

Some styles have changed too, such as with the sliding-out-on-hover
contents box.