grammatical change to 1.1, re-addition of notes
I have added an apostrophe to 1.1. Big news.

I have also added a brief version of ext/1.2-communication-in-business,
as the actual doc was lost in some kind of NTFS-on-a-USB-stick disaster.
It should be enough to remind me what to right for the assignment at
least. In the next commit, I think I'll properly remove all Windows line
endings.