unix line endings, 1.2 and 30.1 work
I have grepped the repository for all files containing Windows newlines
and have converted them to using Unix lines endings with Vim. I know not
whether Sublime Text 2 running on Windows will or is even able to
respect this and keep it the same when saving, but if not, I will simply
copy changes manually to the main repository, and commit there after
checking for any Satan-esque line endings. I've said this in commit
messages about three times, so I'll shut it now.

I have also worked on the 1.2 assignment, as well as the 30.1 one. There
is still much to do on these, as there is with 6.1, however.