blockquote styling, 1.2 work, permalink markup
I've changed the styling for headings and made them somewhat acceptable.
There's less bloat I think, and they look better in practice. Despite
this, I am considering changing them again so that the quotation marks
are at the sides of the content, and not above and below. I might even
use an SVG for it, as it would make life infinitely easier, and they
wouldn't need to be Source Sans quotation marks.

In the headings handler class, I changed the line that replaces markdown
headings with HTML ones, so that the permalink wouldn't have text in.
This means a slight loss in accessibility (how many screen readers will
ever see this though?) but makes life much easier when copying and
pasting into Word.

I have added a bunch to the 1.2 assignment as part of lesson work. I
have also added the "English Grammar" extended document, as required by
an utterly pointless lesson exercise.