contents styling, added .py for 6.1, newlines,..
I have changed the styles for the contents list on the left of the page,
making it fit better in all sorts of viewports. I essentially changed to
using static values in terms of pixel rather than some weird percentage
thing I'd written before.

I also worked on the 6.1 assignment, or at least had the intention of
doing so. I added a line to the assignment to include the file, and
worked a lot on the part of index.php that handles code includes.
Overall, it's a lot better now, with proper tab-indentation support
using modulo and all, and no need for code includes to be pre-indented.

I have also made sure a bunch of files have a trailing newline line in true,
GNU fashion. Similarly, I converted the SVG icon image to Unix newlines,
which hadn't been done for some reason.

I have also added an ext document as part of Unit 1 classwork. You guessed
it; it's another of those pointless ones. This is what I'm supposed to use
OneNote for.

I mentioned it before, but index.php has changed a lot. I have reworked
a lot of the code that handles includes. Overall it's neater and more
consistent, and has more functionality.

functions.php changed slightly, in that I removed the line that would
remove any string of spaces longer than two. This was more of a pain in
the neck than anything.