finished 30.1, contents styling, code includes
I have finished 30.1 to the point of submitting it, as I had a deadline
to meet, although one could definitely work on it more. They say
resubmissions aren't so much of thing for the second round of units, so
with a spot of luck I won't even need to finish it. Yes, I am lazy, but
this work it total bull.

I have changed the stylesheet a little, so that the contents are a
little better spaced out on the left.

I have also changed index.php, so that code shows up correctly. This
essentially involved indenting everything with preg_replace(). I may
switch to using a slightly clunkier method with str_replace() for the
performance benefit, but I don't know how much of an issue this is, TBH.