added 1.3 + 30.2, added 13 images, new styles
I have continued work on the 1.3 and 30.1 assignments. This has involved
adding images to the repo; three for the former and 10 for the latter.
I have compressed the 10 JPEGs with Photoshop and jpegoptim, but I
haven't run the SVGs through any optimiser. I'm sure doing so would have
little effect.

Some of the styles have changed too, as I had a random property that was
screwing over side-by-side image pairs.