fair load of 1.3, svgs optimised, new image,..
I have added a good few words to the 1.3 assignment. The second part of
the assignment is due tomorrow, but we'll have time in the lesson, I
believe. I have some more to do, but not exactly loads.

As part of 1.3, I added a little C hello world, as an example of a
program that could be compiled on several OS' and instruction sets. It
was totally unnecessary, but I like code and an excuse is an excuse. I
added a PNG, optimised, as well, showing a screenshot of CLion open.
What a piece of software. I also optimised the three SVGs I added
previously, stripping out all the Illustrator rubbish mostly.

In the stylesheet, I have moved the "text-align" property so that it
only applies to the section element, and not the aside too. This was
causing some weird-looking alignment.

The index.php file has changed too, as the tab-to-sapce indentation
conversion wasn't working for the C source code, which prints a newline
at one point. This has been fixed. I also added a little bit to the
include regex line, so that it picks up .c files.