30.2 work, some 1.3 work, 6 images, pptx file,..
I have done a fair bit of work for the 30.1 assignment, adding six
progress images for the design of the game logo. These images have been
compressed to an unholy degree. I have added pretty detailed
descriptions of the images and the process of creation. I also need to
explain the tools that I used in the process, which might call for a
screenshot of the Photoshop toolbar. I also added some quick notes on
the tools I used that I should cover.

I also worked on the 1.3 assignment, splitting the report out into it's
own file, so that corrections can be seen. This is for one of those
mindbendingly stupid tasks. <3 BTEC! I pleasantly discovered that the
code which processes includes will actually process includes in those
includes, and presumably in those. Yay for nesting!

As part of the 1.3 work, I also downlaoded the presentation made on
Google Slides as a .pptx presentation as the markers are still bound in
the shackles of Microsoft, and daren't even venture to the world of web
applications. As an added bonus, my name isn't shown quite so blatantly.