1.3 work, image for 30.2, highlighting styles
I have done some more work on 1.3, and lost a proportionate amount of my
will to live as a result. Part of this involved adding styles to the
main stylesheet which give "hb"-class elements a blue background. This
is to highlight text I have added during proofreading for the 1.3
assignment. It'll likely see no use anywhere else, as my assumption is
that no other unit is quite as.. unique, shall we say?

Anyway, I also added an annotated image of the Photoshop tool bar as
mentioned before, and have optipng'd it as usual. I have yet to add it
into the assignment, but I will do so tomorrow.