1.3 completed, styles, renamings, index.php,..
This commit was supposed to be made earlier, but I had to commit in
order to submit the 1.3 assignment which was due in a matter of minutes.
making this commit will undoubetly cause me a load of trouble, but oh
well.

The 1.3 assignment has obviously be finshed, and a few things have
happened as part of this. I have added a few files to keep the proofread
repost separate from the original, and the same sort-of goes for the
magazine article. I renamed the hellow world C program so that it fits
my naming convention properly.

The stylesheet has changed to include Droid Sans as I decided to set the
article in a serif. I should probably include Montserrat as well, for
the sake of completion, but I haven't yet. Some part of me seems to
think that loading four webfonts is somehow several magnitudes more
scandalous than loading three.

I've added a laod of styling for the article other than the typography
as well. There's stuff for the horizontal rules, headings, images and
columns of text. I moved a couple of other blocks around as I was graced
with the unchanged-CSS-suddenly-not-working-as-it-was-before curse.

The headings-handler class has changed so that the headings it converts
from markdown into HTML do not have full permalinks. Perhaps they aren't
strictly speaking permalinks anymore, but I think most browsers should
handle a "Copy link" type thing fine. I might add back full addresses
when I have a little more time.

I renamed the pardown class and updated the index.php include line so
that it followed the naming I was using for the headings-handler class.
I also moved a bunch of code from the index.php file into a separate
function in the functions.php file. This means it's able to process
includes inside includes in the markdown documents. I corrected some of
the comments in the PHP files too.

I think that's it. Fun fun fun.