"finished" 30.2 assignment, renamed + added images
I have finished the 30.2 assignment, or as much as I can without running
out of time or losing the will to continue with life. I have done all of
the assignment other than the distinction task at the end. I may be
given the chance to resubmit, but I think I have to have attempted it
first, which I haven't. I'm not overly fussed anymore.

As part of the assignment, I added two new images, including a full size
poster and an 800px-wide version, both with the corrections suggested by
"classmates". I have also renamed a load of the images, but doing so was
a waste of time more than anything. At least "git mv" is pretty
engrained in my now.