1.4, 6.2, 30.3, 4 images, new table styles
I have worked on the fourth Unit 1 assignment in lessons, which is
looking up, I'd say. We're to set ourselves targets to complete in three
weeks, which for me will be the completion of Functional Stoneware. As
part of the assignment, I added new styles for making smaller tables,
much in the same way as right- and left-floated images are.

I have also done some more work on the 6.1 assignment, as it turns out
nobody understood the assignment. Sounds like it, perhaps, wasn't worded
right? No? Just me and my conspiracies then. I added an SVG flowchart
for this too.

Some work has been done on the 30.3 assignment too. This is all about
the legalities of digital design, which should be well fun. I have
started talking about the sources of imagery for the poster created in
the second assignment. I added three thumbnail/preview images for this.