1.4 work, the big renaming, started time-line
I have done some, but never enough, work for the 1.4 assignment. It
isn't due for an age, but we were to complete the Belbin test and
Memletics test sections for tomorrow, which I haven't.

I have instead spent time on a mass renaming endeavour. I have
essentially zero-padded a bunch of the file names so that unit 6, for
example, is correctly sorted before unit 30. I still have to change the
names of a lot of image files, but I will do this some other time.
Perhaps I'll waste lesson time at college doing it, trying to make it
work in a Windows environment.

I think I have updated all references to files which have been renamed
but I shall verify this some time. So much to do, eh? I also have to
submit more for the 1.3 assignment because of the collective ineptitude
of BTEC, my lecturer and me.