fixes, 1.4 work, (little) 30.3 work
I have done a good amount of work on the 1.4 assignment, describing the
different Belbin test team roles. I have also done some very minimal
work on the 30.3 assignment. Astoundingly little, in fact, considering
the 3 hours I had in lesson working on it. I also may a quick correction
to the 30.2 assignment for something or other.