1.3 + 1.4 + 6.3 + 30.2 + 30.3 work, https fix,..
I have worked on a bunch of assignments. For the unit 1 pair, I moved
the next-ten-years time-line to a better location, and renamed it; it's
part of assignment 4, not 3. I also wrote a fair amount of the
time-line's content, which was taxing. There's a little more to do, but
I have a nifty tactic to cheat a little. Oh, the power of dittos. The
unit 6 assignment has created, but is little more than headings at the
moment. The 30.2 assignment was corrected and had more added to it, as
we apparently have to show changes based on feedback for the poster AND
logo. For the 30.3 assignment, I added bullets for myself later on, and
a reference to the cloud photo I used in the logo feedback redesign.

I've a bunch of images too. Unit 6 gets an SVG "data flow diagram",
while unit 30 gets 2 new images and 2 updated ones. These haven't been
run through jpegoptim yet, but I doubt doing so would yield much.

I've started work on the templating system, but this isn't quite up and
running yet. This accounts for the changes to functions.php.

I have also added styling for superscript text in the main stylesheet,
and changed the links to web fonts from http:// to just //. If the page
is loaded over HTTP, HTTP will be used. If the page is loaded over HTTPS
the fonts will be too. Unless it's IE6, but let's not think about that.