pimp my repo, 1.4, 6.3, 30.3, time-lines
I have made the switch to Ciconia for my Markdown-processing needs,
rather than Parsedown, which I was using before. It has already proven
to be more aware of the Markdown syntax, and has solved some odd issues
I was having with lists. As part of this transition, i have added
Composer to the project, which is a PHP package/dependecy manager. This
will allow me to upgrade Ciconia as new versions are released with much
greater ease. It'll also integrate better with Travis CI, which I have
added a config file for. I'll set that up at some point. I have yet to
update the license file to reflect the change from Parsedown to Ciconia.
I have changed the index.php and functions.php files to use Ciconia, and
have added a function that will return a Ciconia object with GitHub
flavour bells and whistles switched on.

Asides from that, other PHP stuff has changed too. I have changed
comments, changed several used of double quotes to uses of single
quotes, and made the include lines more robust.

I have worked on the 1.4 assignment, finishing the time-line and writing
the approriate CSS to make it come alive. I have also added the one pass
task to the 30.3 assignment, which was all that was worth doing. I also
slightly edited the SVG data flow diagram for the 6.3 assignment,
removing an unnecessary arrow.

This is basically another of those commits which should probaboy be
about four, but which has been put off for some time. There have
probably been a load more little tweaks, but I want to sleep, so this is
all I'm writing.