some 1.4 work, class calling changes
I have done some work on the 1.4 assignment, finishing the diary
entries, despite not having completed the objective, and adding headings
for the ~1600 word report I didn't know had to be written.

I also changed the way in which the functions in the HeadingsHandler
class were being called. Rather than creating an unnecessary instance of
the class, the script now just calls it directly with the
Class::method() syntax.