6.3 finished
I have finished the 6.3 assignment, although it has not been proofread.
I'm sure it's littered with errors. I would run it through Word before
committing, but I need to commit it so I can pull changes on the server
and all.

The flow chart SVG has changed; I made the lines properly black, not the
dark grey that Illustrator sometimes outputs. I also made a small change
to the Blackjack Python script in the 6.3 assignment.